+++ src/constants.ts
@@ -1,0 +1,10 @@
+/* commands */
+export const CMD_TOGGLE_QUOTES = 'advanced-code-toolkit.toggle-quotes';
+
+/* parameters */
+export const Q_QUOTES_ORDER = 'toggle-quotes.quotes-order';
+
+/* messages */
+export const M_ERROR = 'error';
+export const M_WARNING = 'warning';
+export const M_INFO = 'info';