+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## [0.1.0] - 2026-05-18
 
+- Added toggle quotes feature
 - Added commitlint
 - Initial release